[driver_test_realm] Use realm_builder_exposed_dir to forward connections
Prior to this change, we were forwarding capabilities individually, but
if we use a directory, we can do them all in one go.  This simplifies
the code with a minor increase in verbosity in the manifest.

This fixes an issue with exposing the fuchsia.hardware.ramdisk service
which wasn't using the forwarding mechanism, and also allows us to
use dtr_exposes rather than the deprecated exposes Start argument.

Bug: b/357779890
Bug: b/358475423
Change-Id: I3442b36b11ef97b402aadb6543bba0658233fb92
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1098232
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Chris Suter <csuter@google.com>