[kernel] Teach IdlePowerThread how to pause the monotonic clock
Adds the functionality needed to pause the monotonic clock to the
IdlePowerThread. All of this code is checked in behind a feature flag
that is disabled, so this CL should have no functional impacts. It will
give us a chance, however, to review the code and get it checked in so
that we can flip the flag as soon as the boot timeline RFC is accepted.

Bug: 341785588
Change-Id: Ia8c5b34c909672977c97ec183e2e5b5ff003fec8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1093967
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Reviewed-by: John Grossman <johngro@google.com>