[netstack3] Introduce rules table in IP state
This change introduces the lock levels and accessor traits for the
top level rules table. Currently we don't install rules from the
Bindings, so the routing decision is still made from the only
default rule in the table which is just using the main table. There
is no change in behavior yet, we rely on the existing tests to make
sure.

Bug: 354724171

Change-Id: I77b1fe99f92fac560166982341ae6e59a02d7184
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1095732
Commit-Queue: Zeling Feng <zeling@google.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>