[reboot] Ignore failure during reboot process
Previously the reboot process attempted to background a reboot  and
immediately exit with a return code of 0. This doesn't seem to work,
however, and a return code of 255 was returned. This change ensures
that no matter the result of the shell command, the script's execution
will continue to eventually get to a successful exit.

Fixed: 322226825
Change-Id: I31e7654d4aa32fb45cbb8117514ea45ae1c53c2d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1099512
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Christopher Johnson <crjohns@google.com>
Fuchsia-Auto-Submit: Christopher Johnson <crjohns@google.com>