[kernel][vm] Avoid SupplyPagseLocked assert misfiring
The assert at the end of VmCowPages::SupplyPagesLocked is overly
aggressive and can spuriously trigger if AddPageLocked fails with
ZX_ERR_NO_MEMORY. This failure causes the loop to 'break' before
incrementing `offset`, leading to a discrepency in the number of
pages pulled from the VmSpliceList, and the number of offsets we
have successfully supplied.

As supplied_len is not used if any other error kind is returned, this
change redefines the function to state this, and relaxes the assertion.

Change-Id: I984625799e468166ad0681055f0dc361b4ddb37f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1099052
Reviewed-by: David Worsham <dworsham@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>