[ffx][log] Remove dependency on FFX daemon_protocol
Refactors the reconnection logic in ffx log to use APIs provided by FFX
to remove the usage of daemon_protocol. This was identified as a blocker
towards the removal of the ffx daemon.

The core change is reasonably simple and was done in Patch#1, the rest
of the change is about refactoring the existing tests to write them on
top of existing ffx APIs that integrate well with the newly introduced
usage of fho::Connector.

I'll make a few additional improvements to the tests in a follow-up.

Multiply: ffx_log_lib_test
Fixed: 299981807
Change-Id: I960d2f56b2766ec07bb4a4dafcde17fc8bf28718
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1098792
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Brian Bosak <bbosak@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>