[ffx][repo]Correctly handle daemon server in list
This adds calling the FIDL proxy to determine  the state of the
repository server when running as part of the ffx daemon. The previous
logic relied on the process status to determine if it is running or not.
The server in the daemon can be stopped without changing the process
status, which causes the list command to incorrectly report the server
is running.

Bug: b/359290928
Change-Id: I3573bac08de2ef5141a0695fe0ab7b10a8351e98
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1099633
Reviewed-by: Thomas Zander <thomasalpinus@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>