[ffx][repo] Moving server instance writing
This moves the writing of the repositoryy server instance info
to be after starqting the server task. Previously, the instance
data would written when the process starts, but would miss writing the
instance data if the server was stopped then restarted within the same
daemon process.

Bug: b/359290928
Change-Id: Ieb37a0ea7d615cab99c245d6929ca4484353824e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1099634
Reviewed-by: Thomas Zander <thomasalpinus@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>