[driver_runtime][rust] support for "messages"
This implements a "message" object that can be used to hold an arena and
accompanying data, and then allows unpacking that data in various ways.
It also adds a handle type for managing driver runtime handles, which
are stored the same as zircon handles, but are slightly different.

Fixed: 359226657
Change-Id: Iaab4cf212949ad1c58206a479952d58266bb30da
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1087878
Commit-Queue: Megan Batty <mgnb@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>