[usb-hid] Use left-shift operator to calculate power of two.
The driver was using the pow() function to calculate the polling
rate. It invoked libc math functions which are restricted in
drivers.

This CL changes the driver to use the << operator to calculate
the polling rate so that it won't call the libc math function
anymore.

Bug: 359565197
Change-Id: I00b45cf42e483feae28896d6bd36a7f2994c2f2b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1100372
Fuchsia-Auto-Submit: Yilong Li <liyl@google.com>
Commit-Queue: Yilong Li <liyl@google.com>
Reviewed-by: Caroline Liu <carolineliu@google.com>