[starnix] Connect NMFS to socketproxy
Add a Network Manager that has a proxy handle to the socket proxy
StarnixNetworks protocol.

The filesystem is the source of truth for network updates, so it will
maintain the state of the networks even when the socket proxy hasn't
successfully received the default/add/update/remove event. This allows
the Manager to perform a re-transmission of network information such as
when the socket proxy component restarts.

Tested: manual tests, added tests with socketproxy present and responses
mocked, fx test starnix-tests --test-filter=*nmfs*

Bug: 339290072
Bug: 348642138
Fixed: 339448114
Change-Id: I2d088604b808f35a8a95016312348451595cffe8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1073037
Reviewed-by: Theodore Dubois <tbodt@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Zoey Bush <zmbush@google.com>
Reviewed-by: Devon H. O'Dell <dhobsd@google.com>