[kernel][pi] Use latched version of `now`
During each of the PI related operations, a current "now" time (on the
monotonic timeline) is latched at the start of the operation and
passed to HandlePiInteractionCommon to be used as the definition of
"now" for the rest of the operation.

However, during ThreadBaseProfileChanged and
UpstreamThreadBaseProfileChanged, the value latched and passed to the
common handler was not the value being used by the callbacks used to
maintain the dynamic scheduling parameters.  Instead, a new value of
`new` was being computed during the callback using
Scheduler::CurrentTime().

Fix this by capturing the value of `now` computed at the start of the
operation in the lambda currently being used for the callbacks for the
base profile changed operations.

Change-Id: I207dbf4586aeed7cc92fe3aae08e8642f6b2279c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1093079
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: John Grossman <johngro@google.com>