[ffx][logging] Allow subtool override of logging file.
This adds a method in FfxMain trait to allow the subtool
to override the basename of the log file to use.  This
then takes effect just prior to running FfxMain::main().

The basename is not changed if the log file is specified on
the command line.

This change also changes `ffx doctor --record` to capture all
.log files in the log dir.

Bug: b/358404932
Change-Id: If511f3ec609559bc32a21e291ed6346c089fe9e2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1099613
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>