[bazel] pass copy_debug_symbols in bazel_action
The debug symbols for product bundles were not being copied into their
final location because we were not actually passing along the
copy_debug_symbols value for the directory_outputs scope. This also was
causing us to not track all values that were being passed in to the
tracked_files field when we did not specify copy_debug_symbols.

This CL does not fully solve the problem of product bundles not copying
debug symbols but that will be fixed in a later CL since it has to do
with how we are exposing the symbols in the bazel build rule.

BUG: 354039207
Change-Id: I1eba14ef3d6563f01525d12a6698a252b994472b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1100414
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>
Reviewed-by: David Turner <digit@google.com>