[mypy] Fast path for local checks.
Due to bug 339343263 we need to disable mypy incremental caching
otherwise we see flakes in infra, but this unfortunately slows down
local iteration fairly significantly (~1s -> ~8s for later files in this
CL chain).

This CL changes mypy checking to default to incremental mode, and if
that fails it will re-run the check without incremental mode to confirm
the errors. Locally this results in an additional ~1s delay when there
is a valid error, but significant time reduction when there is no error.

Bug: 339343263
Change-Id: I925a281db0f329874531bdb65fda6226df1133eb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1100896
Fuchsia-Auto-Submit: Christopher Johnson <crjohns@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Mahesh Saripalli <maheshsr@google.com>
Reviewed-by: Kevin Cho <chok@google.com>