[components] Fold ComponentRunner epitaph into OnStop
https://fxrev.dev/1094214 added an OnStopInfo event to ComponentRunner,
which lets a runner report a component's exit code to component manager.
(The exit code is the status returned by the _program_ for example the
value passed to exit() in a C program.)

The channel's epitaph, renamed to the component's "termination status", is
used for the status code reported by the _runner_, as documented in the
ComponentController fidl doc. Right now the termination status and the
exit code are reported in two different ways: one by an epitaph, the
other by an event.

This changes the behavior so that both the exit code and the termination
status are reported in the OnStopInfo event (renamed to OnStop). If
OnStop is used, the channel epitaph is ignored.

In the future, we could update all runners to adopt OnStop and deprecate
the epitaph, although no effort is made to do so in this patch.

Fixed: 358182549
Change-Id: I595069c3c4dc1975a0f9d9f4d80fc4f61e6c291f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1099940
Reviewed-by: David Gilhooley <dgilhooley@google.com>
API-Review: Gary Bressler <geb@google.com>
Commit-Queue: Gary Bressler <geb@google.com>