[netstack3] Return from process_fragment
... instead of calling a dispatch function.

This makes control flow a little more clear, and will be especially
useful as we move packet reassembly to before we first hit filtering.

This required removing the packet parsing step from fragment reassembly.
It pushes some work onto the caller, but is required in order to work
around lifetime issues introduced by attempting to pass around the
buffer and the packet (which references the buffer).

Bug: 330912179
Change-Id: I8683f6d60f7e11bf473287fc3c34c1b9dd720207
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1099652
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Peter DeLong <pdelong@google.com>