[netstack3] Reassemble packets before filtering
While it's possible to track, filter, and NAT fragments without
reassembly, it's a technical challenge. For the sake of expediency,
we've decided that Netstack3 should reassemble fragments before
filtering, even on the forwarding path.

This solution is not RFC-compliant, but it's also what Linux does
whenever conntrack is enabled, and it doesn't cause major problems.

See https://fxbug.dev/345814518 for our idea of how to avoid this
reassembly step entirely.

Fixed: 330912179
Change-Id: I0dfbad73b6fb4c86f5e46732701e27bd95977843
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1099653
Commit-Queue: Peter DeLong <pdelong@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>