[dwmac] Implement NetworkDevice protocol
Directly implements the NetworkDevice protocol in the dwmac driver
instead of relying on the netdevice-migration adapter.

This is almost universally an improvement in benchmarks, with the
exception of udp/send having higher CPU usage. However, this higher
CPU usage is corrolated with increases in throughput, so it's not
clear this is the driver's fault.

Bug: 42068759
Change-Id: I97faa35e22a9e3bb786820d0a15a4a105001a3ba
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1089712
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Bjoern Johansson <bjoernj@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>