[bazel_sdk] propagate debug symbols to product
This CL will run an aspect to collect all of the FuchsiaDebugSymbolsInfo
providers in a product and expose them as a top level
FuchsiaDebugSymbolsInfo.

The build-id directories are then exposed as an OutputGroupInfo so they
can be copied into the root build-id directory when building in
fuchsia.git.

BUG: 354039207
Change-Id: I9f608fc29b388ef8efb304b5c1672fc6945b8939
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1101273
Commit-Queue: Chase Latta <chaselatta@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>