[audio][adr] First-reply WatchDevicesAdded
Method fuchsia_audio_device/Registry/WatchDevicesAdded is a
normal hanging-get, used to asynchronously notify clients of
device arrivals. Currently it pends until there is at least
one device, even the first time that a client calls this.

Change this to respond immediately to each client's FIRST
call to this method, even when this means returning an empty
vector. This enables clients to safely use the method
synchronously, on systems where there may not be any audio
devices.

In practice, ADR is eagerly started, and all devfs devices
are detected and initialized before the ADR receives the
first `WatchDevicesAdded` call. However, a subsequent CL
will ensure that this is the case even if ADR is not eagerly
started, by changing ADR's detection and initialization code
to wait for all preexisting devices to be fully initialized
before responding to this method.

Bug: b/344006668
Bug: b/359344835

Change-Id: I2bc33632e58582928e1ad2bb881827b812cacce5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1100522
API-Review: Dale Sather <dalesat@google.com>
Reviewed-by: Dale Sather <dalesat@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>