[bt][gap] Integrate LegacyPairingState with PairingStateManager
This CL wires up the legacy pairing process code to the rest of the
system via the PairingStateManager.

Depending on the pairing events we encounter before/after the ACL
connection is complete, or depending on the result of interrogation with
the peer, we will tell the PSM whether to use SSP or legacy pairing.

Bug: b/42173830
Test: fx test //src/connectivity/bluetooth
Manual Test:
- Paired and streamed audio to/from Nelson DUT with iOS and veho devices
  supporting SSP to ensure SSP still functions
- Paired and streamed audio to/from Nelson DUT with iOS and veho
  devices. Cherry-picked fxr/1078858 to simulate legacy pairing

Change-Id: Ibf745d6452b5f8fc53dbd16101f2cc3d6a20c114
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1062293
Reviewed-by: Ben Lawson <benlawson@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>
Commit-Queue: Lulu Wang <luluwang@google.com>