[kernel][vm] Correct old refptr upgrade comments
The refptr upgrade use to be able to fail when pages were cleaned out
of a VmCowPages in the destructor. This is no longer the case and these
couple of upgrade sites have a stale comment that could lead to
confusion.

For additional clarity a check in the evictor for a valid cow is
changed to an assert with an updated comment.

Change-Id: Ib6a5bd93b454261f3281e3016b6ea85fb9a62c35
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1099879
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: David Worsham <dworsham@google.com>
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>