[sysmem][assembly] Move sysmem driver to bootstrap
This change allows the microfuchsia product to avoid including the
sysmem driver. It's possible that sysmem should be moved to it's own AIB
and boards/products should opt into a feature that causes sysmem to be
included, but for now, all products other than microfuchsia need it, so
that can be done at a later point.

Fixed: 335548998
Change-Id: I9a98056b47532ec8f53bb06d76e421ece40a283a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1101919
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>