[power] Add benchmark test for Topological Test Daemon
The Power Topology Testing fidl calls are tested for performances using
criterion. An integration test was created to verify the correctness of
the functions acquiring the fidl proxies, and actually toggling lease
states.

The location of this bench is moved from the System Activity Governor
to src/power. The metric names are revised as well.

Bug: b/357445279
Test: fx set terminal.vim3 --with //src/tests/end_to_end/perf:test \
    --with-test //src/power:tests --release
  fx test --e2e power_framework_microbenchmarks -o
  fx test -o power-framework-bench-integration-tests
    --test-filter=*test_topologytestdaemon_toggle -- --repeat 5000
Change-Id: Idb7967896ac575d4cbbfed3695958c92634bfe85
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1098517
Commit-Queue: Hai Bi <bihai@google.com>
Reviewed-by: Didi She <didis@google.com>
Reviewed-by: Michael Brunson <mbrunson@google.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>