[ffx][fho] Expose generalized daemonize
This exposes starting a command as a daemon process.

This is added to support long running tools such as repository server
to run in the background in the same way the ffx daemon is run in the
background. This includes the libc call to make the daemon process, and
also set up the logging to capture the output of the daemon process.

Bug: b/358390369
Change-Id: Id46e24d5035a85780a1196cdb24d7dd5c9cbdf01
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1101152
Reviewed-by: Andrew Davies <awdavies@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>