[devices][testing] Adds syscall-intercept library
This is a library to be used in power observability
integration tests.

We want to verify that a program running under Starnix
eventually causes the zx_system_suspend_enter syscall.

This library intercepts the syscall and allows us
to gather call count from it.

Bug: 356216355
Change-Id: I31de3b049bdce7f4f459cb8fe8e5cbd7df4502f3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1099410
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>