[ctf] Reland "[ctf] Add component_manager tests to subbuild"
This CL fixes the underlying problem with the test. Including the
component_manager with a config for running in a nested setting also
caused a "data" directory to appear, which is asserted in the test. This
change now adds a placeholder config which results in the data directory
appearing, allowing the existing assertion to pass.

This reverts commit 9093dc6b2d61a538e5c631688317027b704a0bb0.

Run-All-Tests: true

Change-Id: I138cb97b56e66be5163163a22bfe0b4c17dc9db8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1103312
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Brian Bosak <bbosak@google.com>
Fuchsia-Auto-Submit: Christopher Johnson <crjohns@google.com>
Commit-Queue: Christopher Johnson <crjohns@google.com>