[ffx] Update OvernetConnector to exit early.
If a channel closes, it may never occur that other channels feeding into
OvernetConnector close as well. The result of this is that for things
like SSH, if the underlying command stops, not all channels will exit,
meaning users will not get PEER_CLOSED for things they expect (like when
rebooting a device).

For this, all completion of channel reads are instead done as a race
(using future_lite's `or` combinator instead of `zip` operator) instead
of requiring everything must complete at once.

Change-Id: I565f5bb63fa4b3ee635f1544fc3e8f70d1f149c1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1099468
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Andrew Davies <awdavies@google.com>