[bt-transport-uart] Fix crash
A crash happened because a receive packet came before HciTransport
connection is established. This was fixed in the test with a
libsync::Completion, but it also observed on a real device now. Fixing
it with a new trick -- delay the start of read until HciTransport is
available.

This change also fixes the flake in unittest which results from a late
establishment of ScoConnection protocol.

Test: Unittest and on device test.

Change-Id: I7eb2e96cf5dadd07a79bfe4b1a6acb7e12da8607
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1097832
Reviewed-by: Ben Lawson <benlawson@google.com>
Commit-Queue: Zhiyi Chen <zhiyichen@google.com>