[starnix][input] Use unique Fuchsia VID and PIDs
It turns out that implausible USB VID or PID prevent correct
device configuration in Android. Fix that by adopting
plausible-seeming identifiers.

The system is already configured to use the correct files for
keyboard. For touch, this is not the case; but the current state
of touch configuration is broken anyways, so I don't expect
things to change after this change goes in. It does give us
a chance to right the config on Android side, though.

Bug: 353765537
Change-Id: Idfba4f0a83086657f19920d67024d395bfbe0eaa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1102435
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>