[kernel][vm] Improve CSM lock annotations
Changes the ContentSizeManager::Operation to have a constant parent_
references and use a separate valid_ flag for tracking validity. This
allows for the external users of the operation to used AliasedLock
acquisitions instead of having to assert that the lock is held.

Asserting that the lock is held is still needed for when the CSM itself
is iterating over the operation queues.

Change-Id: I2554a9fc7ac89c5168f8e11c7da357ee9991efe1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1103752
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Reviewed-by: David Worsham <dworsham@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>