[kernel][vm] Free pages directly when reclaiming
Changes the contract of all the eviction/reclamation/discard APIs in
VmCowPages to not return ownership of pages to the caller, but rather
directly free them. This has two primary benefits:
 1. Prevents a class of bugs where callers of these methods forget
    that they own the pages and need to free them, leading to memory
    leaks.
 2. Allows for better utilization of the page caches as freeing
    externally was bypassing these caches.

The names of some of these methods is changed to better reflect their
new semantics.

Change-Id: Ia4b88c385d6d6bd8a64b57357d826747f4805b9f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1100856
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: David Worsham <dworsham@google.com>