[starnix][kernel] Integrate timer mod into kernel
The timer mod is too small to be a top-level mod. This CL merges the
timer mode with the reset of the kernel. The abstract TimerFile and the
TimerOps belong in the VFS. The Timeline belongs in the task mod.

Change-Id: I616909a10c95019ec3c15d61b541cf2664444f35
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1103274
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Guocheng Wei <guochengwei@google.com>