[sestarnix] Resolve file system labeling scheme details
Introduce a union expressing the different possible file system
labeling schemes, and resolve the security contexts to use for the
file system, its root node, and any unlabeled nodes, once a policy
has been loaded.

This will be updated to take account of the labeling scheme and
settings specified in the policy for the filesystem type.

Bug: 349800754
Co-authored-by: Wez <wez@google.com>
Change-Id: I31ff401cd7d323af1f68f3043e1970889b3d4670
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1087493
Commit-Queue: Nathaniel Manista <nathaniel@google.com>
Reviewed-by: Wez <wez@google.com>