[dtr] provide add function for dtr_exposes/offers
This provides a function to add the exposes and offers that are being
provided through the dtr_exposes/offers fields. Also adds the feature to
the c++ variant of the realm_builder.

This change also improves the consistency between the routes we add with
the various realm builders and the driver test realm component manifest.

Bug: 358475423
Change-Id: I264d87257474f14edd6150c0a9908b5ddaf8e793
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1098077
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>