[ffx][repo] Adding --background option to serving
This adds the `--background` option to ffx repository server start.

This option starts the standalone "foreground" server as a daemon
process.

There is validation added before starting the daemon to be able to
provide errors on the console vs. having to dig through the log.

Bug: b/358390369
Change-Id: I1d5946c3c020a97e8d47ce33a9d6c49bc5e41dde
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1101192
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Thomas Zander <thomasalpinus@google.com>