[time] Allow force-starting from RTC
Force-starting the clock was so far done only when RTC is not present.

This change does the same if RTC is present, but its reading is known
bad. In which case, we start UTC from backstop, and leave the faulty
state in RTC until the next regular time update.  If the device is not
network-capable or network-connected, that RTC update may never come.

In normal operation this behavior is relevant only when RTC is
unpowered by a battery, or pristine. Both are very unlikely scenarios in
general. We do deal with unpowered RTCs in some devices today, but this
is supposed to be a temporary state of affairs.

Bug: 355293758
Change-Id: Ie1fc0f6b9f93e0638d80e57d0491c52e7170d9a3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1100898
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: Mukesh Agrawal <quiche@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>