[ffx][repository] Add IfThisThenThat lint for devhost literals
For the usages of the default `devhost` repository in ffx, add
IfThisThenThat linting to help developers in case they want to change
this default in the future.
Note that there is a current limitation that the linter can't have
multiple targets in the ThenChange(path) clause.

Bug: b/359534719
Change-Id: I81041f0c60c410c73ae6ea4024934fa98bfe229b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1103656
Fuchsia-Auto-Submit: Thomas Zander <thomasalpinus@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>