Reland "[starnix] Support alarm type in POSIX timer syscalls"
This is a reland of commit 86d065254e1d8a214f4f4f02d59cd31c4067e5f3

The revert was due to flakes in the newly added syscall tests. The root
issue is the remaining time of the timer should be non-negative. There
might be a delay in updating timer states in another kthread. So the
fix is to ensure the remainder of the timer is a non-negative number.

Original change's description:
> [starnix] Support alarm type in POSIX timer syscalls
>
> Test: fx test ltp_tests_bundle_t --test-filter="timer*"
> Test: fx test test timers_test
> Bug: 349190823, 349191846
> Change-Id: I3df48fe662aa4db7677d548b2ad85385e64b49c0
> Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1076646
> Commit-Queue: Guocheng Wei <guochengwei@google.com>
> Reviewed-by: Adam Perry <adamperry@google.com>

Bug: 349190823, 349191846
Multiply: fuchsia-pkg://fuchsia.com/starnix_syscalls_cpp_tests#meta/timers_test.cm: 20
Change-Id: I3fedba5e5b33068468cc0027465bca6ba1eb8f84
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1104312
Commit-Queue: Guocheng Wei <guochengwei@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>